requirements.txt
crewai 
crewai-tools
streamlit
Pillow
opencv-python
pytesseract
